requirements/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements/requirements.txt
awscli==1.32.97
    # via cladetime (pyproject.toml)
boto3==1.34.97
    # via cladetime (pyproject.toml)
botocore==1.34.97
    # via
    #   awscli
    #   boto3
    #   s3transfer
certifi==2024.2.2
    # via requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   cladetime (pyproject.toml)
    #   rich-click
cloudpathlib==0.18.1
    # via cladetime (pyproject.toml)
colorama==0.4.6
    # via awscli
docutils==0.16
    # via awscli
idna==3.7
    # via requests
jellyfish==1.1.0
    # via us
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
markdown-it-py==3.0.0
    # via rich
mdurl==0.1.2
    # via markdown-it-py
numpy==1.26.4
    # via
    #   pandas
    #   pyarrow
pandas==2.2.2
    # via cladetime (pyproject.toml)
polars==1.6.0
    # via cladetime (pyproject.toml)
pyarrow==16.0.0
    # via cladetime (pyproject.toml)
pyasn1==0.6.0
    # via rsa
pygments==2.18.0
    # via rich
python-dateutil==2.9.0.post0
    # via
    #   botocore
    #   pandas
pytz==2024.1
    # via pandas
pyyaml==6.0.1
    # via awscli
requests==2.32.3
    # via cladetime (pyproject.toml)
rich==13.7.1
    # via
    #   cladetime (pyproject.toml)
    #   rich-click
rich-click==1.8.0
    # via cladetime (pyproject.toml)
rsa==4.7.2
    # via awscli
s3transfer==0.10.1
    # via
    #   awscli
    #   boto3
six==1.16.0
    # via python-dateutil
structlog==24.1.0
    # via cladetime (pyproject.toml)
typing-extensions==4.11.0
    # via rich-click
tzdata==2024.1
    # via pandas
urllib3==2.2.1
    # via
    #   cladetime (pyproject.toml)
    #   botocore
    #   requests
us==3.2.0
    # via cladetime (pyproject.toml)
